--- src/utils/validators/validateDuration.ts
@@ -1,9 +1,0 @@
-export const validateDuration = (duration: any) => {
-  if ((!duration && duration !== 0) || !Number.isInteger(+duration)) {
-    return "Duration must be an integer";
-  }
-  if (+duration < 0) {
-    return "Duration cannot be negative";
-  }
-  return "";
-};
