--- src/utils/validators/validateEmail.ts
@@ -1,1 +1,0 @@
-export const validateEmail = (v: string): boolean => /\S+@\S+\.\S+/.test(v);
