--- src/utils/validators/validateJira.ts
@@ -1,1 +1,0 @@
-export const validateJira = (v: string) => v.match(".+-[0-9]+") !== null;
