--- src/utils/validators/validatePercentage.ts
@@ -1,9 +1,0 @@
-export const validatePercentage = (percent: any) => {
-  if (!Number.isFinite(percent)) {
-    return "Percent must be a number";
-  }
-  if (+percent <= 0) {
-    return "Percent cannot be negative";
-  }
-  return "";
-};
