--- src/utils/validators/validateSlack.ts
@@ -1,2 +1,0 @@
-export const validateSlack = (v: string): boolean =>
-  v.match("(#|@).+") !== null;
